# openapi.yaml
openapi: 3.1.0
info:
  title: Jinbe API
  description: |-
    Jinbe API for Kubernetes cluster and database management.

    ## Authentication
    Most endpoints require authentication via the `ory_kratos_session` cookie.
    Use the **Authorize** button to set your session cookie value.

    Public endpoints (no auth required):
    - `GET /api/health`
    - `GET /api/whoami`
  version: 1.0.0
  contact:
    name: Jinbe Team
components:
  securitySchemes:
    kratosSession:
      type: apiKey
      in: cookie
      name: ory_kratos_session
      description: Ory Kratos session cookie. Get this from your browser after logging in.
...
paths:
...
  /api/admin/rbac/bundle/export:
    get:
      tags:
        - rbac
        - backup
      description: Export RBAC config as a portable JSON bundle.
        ?sections=services,groups,… narrows it; omitted = full 1:1 snapshot.
      parameters:
        - schema:
            type: string
          in: query
          name: sections
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/admin/rbac/bundle/import:
    post:
      tags:
        - rbac
        - backup
      description: Import an auth bundle — restores RBAC config. Body must be a full
        snapshot; ?sections=services,groups,… applies only those parts
        (override/add, no prune), omitted = full 1:1 restore.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      parameters:
        - schema:
            type: string
          in: query
          name: sections
          required: false
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  imported:
                    type: object
                    additionalProperties: true
...
  /api/admin/rbac/bundle/backups/restore:
    post:
      tags:
        - rbac
        - backup
      description: Restore RBAC config from an S3 backup snapshot (full replace).
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - key
              properties:
                key:
                  type: string
        required: true
      responses:
        "200":
          description: Default Response
...
  /api/admin/rbac/bundle/history/{id}/rollback:
    post:
      tags:
        - rbac
        - backup
      description: Restore the RBAC config snapshot of a history entry (full replace).
        The current state is snapshotted first (reason 'pre-rollback'), so a
        rollback is itself reversible.
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rolledBackTo:
                    type: object
                    additionalProperties: true
                  imported:
                    type: object
                    additionalProperties: true
...
servers:
  - url: http://localhost:3000
    description: Local Development
security:
  - kratosSession: []
...